+++ frontend-ios/microblog/microblog/timeline/CreateBlogPostViewController.h
@@ -1,10 +1,11 @@
 //  Copyright (c) 2014 lovelysystems.
 
 #import <UIKit/UIKit.h>
+#import "LoginViewController.h"
 
 @protocol CreateBlogPostDelegate;
 
-@interface CreateBlogPostViewController : UIViewController <UITextViewDelegate>
+@interface CreateBlogPostViewController : UIViewController <UITextViewDelegate, LoginViewControllerDelegate>
 
 @property (nonatomic, weak) id<CreateBlogPostDelegate> delegate;